Split out Google API request
