+++ app/services/distance_calculator_service/google_api.rb
@@ -1,0 +1,58 @@
+class DistanceCalculatorService
+  class GoogleApi
+    GOOGLE_DIRECTIONS_API = 'https://maps.google.com/maps/api/directions/json'.freeze
+
+    def initialize(origin, destination)
+      @origin = origin
+      @destination = destination
+    end
+
+    def distances
+      @distances ||= routes.map { |route| sum_legs route['legs'] }.select(&:positive?)
+    end
+
+    private
+
+    def sum_legs(legs)
+      return 0 if legs.nil?
+
+      legs.map { |leg| leg.dig('distance', 'value') }.compact.sum
+    end
+
+    def routes
+      @routes ||= begin
+        JSON.parse(directions)['routes']
+      rescue StandardError => e
+        log("Failed to calculating distance form #{@origin} to #{@destination}", error: e, level: :error)
+        []
+      end
+    end
+
+    def directions
+      @directions ||= RestClient.get(GOOGLE_DIRECTIONS_API, params: params).body
+    end
+
+    def params
+      @params ||= {
+        region: 'uk',
+        alternatives: true,
+        key: Rails.application.secrets.google_api_key,
+        origin: @origin,
+        destination: @destination
+      }
+    end
+
+    def log(message, error: nil, level: :info)
+      LogStuff.send(
+        level,
+        class: 'DistanceCalculatorService::GoogleApi',
+        action: 'routes',
+        origin: @origin,
+        destination: @destination,
+        error: error ? "#{error.class} - #{error.message}" : 'false'
+      ) do
+        message
+      end
+    end
+  end
+end
